Extract Cell class.
